backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch==2.3.1
torchvision==0.18.1
pillow==10.2.0
celery==5.3.6
redis==5.0.1
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
numpy==1.26.4
ultralytics
tqdm
transformers
diffusers
accelerate
peft
open-clip-torch
timm
lpips
dominate
opencv-python
